backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
firebase-admin
pydantic
python-dotenv
